Move IsReadmeFile*() from modules/markup to routers/.../view
These functions don't examine contents, but just filenames
so they don't fit in well in a markup module. view.go is the
only place these functions are ever called, so put them there.
